+++ Information_System_MVC/Controllers/WorkPlaceController.cs
@@ -16,40 +16,49 @@ namespace Information_System_MVC.Controllers
         [Authorize]
         public ActionResult Index()
         {
-            if ((System.Web.HttpContext.Current.Session["CurrentUser"] as ConnectedWorker).Power == 2
-                || (System.Web.HttpContext.Current.Session["CurrentUser"] as ConnectedWorker).Power == 0)
+            if (System.Web.HttpContext.Current.Session["CurrentUser"] is ConnectedWorker)
             {
-                IEnumerable<WorkPlace> workPlaces = db.WorkPlaces;
+                if ((System.Web.HttpContext.Current.Session["CurrentUser"] as ConnectedWorker).Power == 2
+                || (System.Web.HttpContext.Current.Session["CurrentUser"] as ConnectedWorker).Power == 0)
+                {
+                    IEnumerable<WorkPlace> workPlaces = db.WorkPlaces;
 
-                ViewBag.WorkPlaces = workPlaces;
+                    ViewBag.WorkPlaces = workPlaces;
 
-                return View();
+                    return View();
+                }
+                else
+                    return Redirect("/Home/Index");
             }
             else
                 return Redirect("/Home/Index");
         }
 
-
         [Authorize]
         [HttpGet]
         public ActionResult Details(int? id)
         {
-            if ((System.Web.HttpContext.Current.Session["CurrentUser"] as ConnectedWorker).Power == 2
-               || (System.Web.HttpContext.Current.Session["CurrentUser"] as ConnectedWorker).Power == 0)
+            if (System.Web.HttpContext.Current.Session["CurrentUser"] is ConnectedWorker)
             {
-                if (id == null)
+                if ((System.Web.HttpContext.Current.Session["CurrentUser"] as ConnectedWorker).Power == 2
+               || (System.Web.HttpContext.Current.Session["CurrentUser"] as ConnectedWorker).Power == 0)
                 {
+                    if (id == null)
+                    {
+                        return HttpNotFound();
+                    }
+
+                    WorkPlace workPlace = db.WorkPlaces.Find(id);
+
+                    if (workPlace != null)
+                    {
+                        return View(workPlace);
+                    }
+
                     return HttpNotFound();
                 }
-
-                WorkPlace workPlace = db.WorkPlaces.Find(id);
-
-                if (workPlace != null)
-                {
-                    return View(workPlace);
-                }
-
-                return HttpNotFound();
+                else
+                    return Redirect("/Home/Index");
             }
             else
                 return Redirect("/Home/Index");
@@ -60,75 +69,94 @@ namespace Information_System_MVC.Controllers
         [HttpGet]
         public ActionResult Create()
         {
-            if ((System.Web.HttpContext.Current.Session["CurrentUser"] as ConnectedWorker).Power == 2)
-                return View();
+            if (System.Web.HttpContext.Current.Session["CurrentUser"] is ConnectedWorker)
+            {
+                if ((System.Web.HttpContext.Current.Session["CurrentUser"] as ConnectedWorker).Power == 2)
+                    return View();
+                else
+                    return Redirect("/Home/Index");
+            }
             else
                 return Redirect("/Home/Index");
         }
-
 
         [Authorize]
         [HttpPost]
         public ActionResult Create([Bind(Include = "Id,PlaceId,BuildingId")] WorkPlace workPlace)
         {
-            if ((System.Web.HttpContext.Current.Session["CurrentUser"] as ConnectedWorker).Power == 2)
+            if (System.Web.HttpContext.Current.Session["CurrentUser"] is ConnectedWorker)
             {
-                try
+                if ((System.Web.HttpContext.Current.Session["CurrentUser"] as ConnectedWorker).Power == 2)
                 {
-                    db.WorkPlaces.Add(workPlace);
-                    db.SaveChanges();
+                    try
+                    {
+                        db.WorkPlaces.Add(workPlace);
+                        db.SaveChanges();
 
-                    return RedirectToAction("Index");
+                        return RedirectToAction("Index");
+                    }
+                    catch
+                    {
+                        return View();
+                    }
                 }
-                catch
-                {
-                    return View();
-                }
+                else
+                    return Redirect("/Home/Index");
             }
             else
                 return Redirect("/Home/Index");
         }
-
 
         [Authorize]
         [HttpGet]
         public ActionResult Edit(int? id)
         {
-            if ((System.Web.HttpContext.Current.Session["CurrentUser"] as ConnectedWorker).Power == 2)
+            if (System.Web.HttpContext.Current.Session["CurrentUser"] is ConnectedWorker)
             {
-                if (id == null)
+                if ((System.Web.HttpContext.Current.Session["CurrentUser"] as ConnectedWorker).Power == 2)
                 {
+                    if (id == null)
+                    {
+                        return HttpNotFound();
+                    }
+
+                    WorkPlace workPlace = db.WorkPlaces.Find(id);
+
+                    if (workPlace != null)
+                    {
+                        return View(workPlace);
+                    }
+
                     return HttpNotFound();
                 }
-
-                WorkPlace workPlace = db.WorkPlaces.Find(id);
-
-                if (workPlace != null)
-                {
-                    return View(workPlace);
-                }
-
-                return HttpNotFound();
+                else
+                    return Redirect("/Home/Index");
             }
             else
                 return Redirect("/Home/Index");
         }
+
         [Authorize]
         [HttpPost]
         public ActionResult Edit(WorkPlace workPlace)
         {
-            if ((System.Web.HttpContext.Current.Session["CurrentUser"] as ConnectedWorker).Power == 2)
+            if (System.Web.HttpContext.Current.Session["CurrentUser"] is ConnectedWorker)
             {
-                try
+                if ((System.Web.HttpContext.Current.Session["CurrentUser"] as ConnectedWorker).Power == 2)
                 {
-                    db.Entry(workPlace).State = EntityState.Modified;
-                    db.SaveChanges();
-                    return RedirectToAction("Index");
+                    try
+                    {
+                        db.Entry(workPlace).State = EntityState.Modified;
+                        db.SaveChanges();
+                        return RedirectToAction("Index");
+                    }
+                    catch
+                    {
+                        return View();
+                    }
                 }
-                catch
-                {
-                    return View();
-                }
+                else
+                    return Redirect("/Home/Index");
             }
             else
                 return Redirect("/Home/Index");
@@ -138,42 +166,53 @@ namespace Information_System_MVC.Controllers
         [HttpGet]
         public ActionResult Delete(int id)
         {
-            if ((System.Web.HttpContext.Current.Session["CurrentUser"] as ConnectedWorker).Power == 2)
+            if (System.Web.HttpContext.Current.Session["CurrentUser"] is ConnectedWorker)
             {
-                WorkPlace workPlace = db.WorkPlaces.Find(id);
-
-                if (workPlace == null)
-                {
-                    return HttpNotFound();
-                }
-
-                return View(workPlace);
-            }
-            else
-                return Redirect("/Home/Index");
-        }
-        [Authorize]
-        [HttpPost, ActionName("Delete")]
-        public ActionResult DeleteConfirmed(int id)
-        {
-            if ((System.Web.HttpContext.Current.Session["CurrentUser"] as ConnectedWorker).Power == 2)
-            {
-                try
+                if ((System.Web.HttpContext.Current.Session["CurrentUser"] as ConnectedWorker).Power == 2)
                 {
                     WorkPlace workPlace = db.WorkPlaces.Find(id);
+
                     if (workPlace == null)
                     {
                         return HttpNotFound();
                     }
 
-                    db.WorkPlaces.Remove(workPlace);
-                    db.SaveChanges();
-                    return RedirectToAction("Index");
+                    return View(workPlace);
                 }
-                catch
+                else
+                    return Redirect("/Home/Index");
+            }
+            else
+                return Redirect("/Home/Index");
+        }
+
+        [Authorize]
+        [HttpPost, ActionName("Delete")]
+        public ActionResult DeleteConfirmed(int id)
+        {
+            if (System.Web.HttpContext.Current.Session["CurrentUser"] is ConnectedWorker)
+            {
+                if ((System.Web.HttpContext.Current.Session["CurrentUser"] as ConnectedWorker).Power == 2)
                 {
-                    return View();
+                    try
+                    {
+                        WorkPlace workPlace = db.WorkPlaces.Find(id);
+                        if (workPlace == null)
+                        {
+                            return HttpNotFound();
+                        }
+
+                        db.WorkPlaces.Remove(workPlace);
+                        db.SaveChanges();
+                        return RedirectToAction("Index");
+                    }
+                    catch
+                    {
+                        return View();
+                    }
                 }
+                else
+                    return Redirect("/Home/Index");
             }
             else
                 return Redirect("/Home/Index");
